Draft: split mirror from Draft.py
.
